requirements.txt
mixpeek==0.6.8
pydantic==2.6.4
pymongo==4.6.3

openai==1.16.2
gradio==4.25.0
langchain==0.1.14
langchain-community==0.0.31

langchain_openai
langchain_core
python-dotenv
langsmith
langchain-nomic
unstructured